Tagging NetHack 3.6.1 Release

+++ win/win32/mhfont.c
@@ -206,11 +206,11 @@ mswin_charset()
 {
     CHARSETINFO cis;
     if (SYMHANDLING(H_IBM))
-        if (TranslateCharsetInfo((DWORD *) GetOEMCP(), &cis, TCI_SRCCODEPAGE))
+        if (TranslateCharsetInfo((DWORD *) (uintptr_t) GetOEMCP(), &cis, TCI_SRCCODEPAGE))
             return cis.ciCharset;
         else
             return OEM_CHARSET;
-    else if (TranslateCharsetInfo((DWORD *) GetACP(), &cis, TCI_SRCCODEPAGE))
+    else if (TranslateCharsetInfo((DWORD *) (uintptr_t) GetACP(), &cis, TCI_SRCCODEPAGE))
         return cis.ciCharset;
     else
         return ANSI_CHARSET;
